Split tee profile worker tests by concern (#57)
Separate core worker execution, selector concurrency, and CLI validation
coverage into focused test modules. Move the Syrupy snapshots to the
matching module snapshot files so the existing snapshot assertions keep
using the same keys after collection.
